requirements.txt
Flask==2.1.2
Flask-login==0.6.1
Flask-WTF==1.0.1
Jinja2==3.1.2
Pillow==9.1.0
SQLAlchemy==1.4.36
WTForms==3.0.1
Werkzeug==2.1.2
pip==22.0.4
setuptools==56.0.0